Added: record-route supports prefix-based tracing (in this case, exact-matching lookups are performed)
Added: "bgp domain X full-mesh" generates a full-mesh of iBGP sessions in the given domain X. Note that no check is currently performed and this feature is considered experimental (it will only be included in the build if the __EXPERIMENTAL__ symbol is defined).

Fixed: since the addition of route prefix in the net-route-info structure, "show rt" displayed the route prefixes twice. This is fixed.